requirements.txt
plyer==2.1.0
pyshadow==0.0.5
selenium==4.9.1
python-dotenv==1.0.1